+++ src/pdf_edit_lib/pdffile.cpp
@@ -55,57 +55,63 @@ int InputPdfFile::output_page_count()
     return m_output_page_count;
 }
 
-Problems InputPdfFile::set_pages_filter_from_string(const std::string &str)
+void InputPdfFile::set_pages_filter_from_string(const std::string &str)
 {
+    m_interval_errors.clear();
+    m_interval_warnings.clear();
+
+    m_pages_filter_string = str;
+
     m_filters.clear();
     m_output_page_count = 0;
 
-    Problems problems;
-
+    // Invalid characters
     if (str.find_first_not_of("0123456789- ,") != std::string::npos)
     {
-        problems.add(new Problem(
-                              ProblemType::error_invalid_char,
+        m_interval_errors.push_back(IntervalIssue(
+                              IntervalIssue::error_invalid_char,
                               str.substr(str.find_first_not_of("0123456789- ,"), 1))
                           );
-        return problems;
+        return;
     }
 
+     // Void string
     if (str.find_first_not_of("- ,") == std::string::npos)
     {
         m_output_page_count = this->page_count();
-        return problems; //void str
+        return;
     }
 
-    //parse str
+    // Parse string
     std::string::size_type cursor = str.find_first_not_of(" ,-");
     std::string::size_type interval_end = str.find_first_of(" ,", cursor);
 
     while (cursor < str.length())
     {
-        //single number
+        // Single number
         if (str.find_first_of('-', cursor) >= interval_end)
         {
             std::string page_number = str.substr(cursor, interval_end - cursor);
             int num = std::stoi(page_number);
-            Problem *problem = this->add_pages_filter(num, num);
-
-            if (problem != NULL)
-                problems.add(problem);
+            this->add_pages_filter(num, num);
 
             cursor = str.find_first_not_of(" ,-", interval_end);
             interval_end = str.find_first_of(" ,", cursor);
         }
-        //interval
+        // Interval
         else
         {
             std::string::size_type first_number_end = str.find_first_of('-', cursor);
             std::string::size_type second_number_start = str.find_first_not_of('-', first_number_end);
-            if (second_number_start >= interval_end || //syntax error: no second number
-                    str.find_first_of('-', second_number_start) < interval_end) //syntax error: more '-' in one interval
+            if (
+                    // Syntax error: no second number
+                    second_number_start >= interval_end ||
+                    // Syntax error: more '-' in one interval
+                    str.find_first_of('-', second_number_start) < interval_end
+                    )
             {
-                problems.add(new Problem(
-                                      ProblemType::error_invalid_interval,
+                m_interval_errors.push_back(IntervalIssue(
+                                      IntervalIssue::error_invalid_interval,
                                       str.substr(cursor, interval_end - cursor))
                                   );
 
@@ -116,58 +122,75 @@ Problems InputPdfFile::set_pages_filter_from_string(const std::string &str)
             {
                 int from = std::stoi(str.substr(cursor, first_number_end - cursor));
                 int to = std::stoi(str.substr(second_number_start, interval_end));
-                Problem *problem = this->add_pages_filter(from, to);
-
-                if (problem != NULL)
-                    problems.add(problem);
+                this->add_pages_filter(from, to);
 
                 cursor = str.find_first_not_of(" ,-", interval_end);
                 interval_end = str.find_first_of(" ,", cursor);
             }
         }
     }
-
-    return problems;
 }
 
-Problem *InputPdfFile::add_pages_filter(int from, int to)
+const std::string &InputPdfFile::pages_filter_string()
 {
-    //check interval boundaries
+    return m_pages_filter_string;
+}
+
+void InputPdfFile::add_pages_filter(int from, int to)
+{
+    // Check interval boundaries
     if (from < 1 || from > this->page_count() ||
             to < 1 || to > this->page_count())
-        return new Problem(
-                    ProblemType::error_page_out_of_range,
+    {
+        m_interval_errors.push_back(IntervalIssue(
+                    IntervalIssue::error_page_out_of_range,
                     std::to_string(from) + "-" + std::to_string(to)
-                    );
+                    ));
+        return;
+    }
 
     if (from > to)
-        return new Problem(
-                    ProblemType::error_invalid_interval,
+    {
+        m_interval_errors.push_back(IntervalIssue(
+                    IntervalIssue::error_invalid_interval,
                     std::to_string(from) + "-" + std::to_string(to)
-                    );
+                    ));
+        return;
+    }
 
-    Problem *problem = NULL;
-
-    //check if the new interval intersects old ones
+    // Check if the new interval intersects old ones
     std::list<std::pair<int, int>>::iterator it;
     for (it=m_filters.begin(); it != m_filters.end(); ++it)
     {
-        //intersection found
+        // Intersection found
         if (!(from > it->second || to < it->first))
-            problem = new Problem(
-                        ProblemType::warning_overlapping_interval,
+            m_interval_warnings.push_back(IntervalIssue(
+                        IntervalIssue::warning_overlapping_interval,
                         std::to_string(from) + "-" + std::to_string(to)
-                        );
+                        ));
     }
 
     m_filters.push_back(std::pair<int, int>(from, to));
 
     m_output_page_count += to - from + 1;
+}
 
-    return problem;
+const std::vector<IntervalIssue> &InputPdfFile::pages_filter_errors()
+{
+    return m_interval_errors;
+}
+
+const std::vector<IntervalIssue> &InputPdfFile::pages_filter_warnings()
+{
+    return m_interval_warnings;
 }
 
 void InputPdfFile::set_rotation(int rotation)
 {
     m_rotation = rotation;
+}
+
+int InputPdfFile::rotation()
+{
+    return m_rotation;
 }